Fix double free bug for Capp.offline on an IO
Both the ruby IO and capp close the IO when garbage collected.  This will
lead to a double-free and possible segmantation fault.

Now capp dups the file descriptor and creates a new stream (FILE *) for the
file descriptor that libpcap can safely close without closing the ruby stream
(and vice-versa).