Work around rake-compiler trying to copy .gemtest
Hoe adds .gemtest to the files list automatically for packaging.  This
file only exists in the package directory.  rake-compiler wants to copy
this file but it does not exist when trying to do this.

This might remove .gemtest from the package, but it seems the
functionality it provides is currently dead.